requirements.txt
Flask>=2.2
psycopg2-binary>=2.9
python-dotenv>=1.0
Flask-WTF>=1.1
gunicorn>=20.1
psycopg2>=2.9
